feat: School 모델 점수 필드 변경, 점수 계산 로직 변경
- School 모델의 school_score 필드를 BigIntegerField에서 DecimalField로 변경하여 점수 정밀도 향상
- ELO 점수 계산 로직에서 점수 차이를 반영하도록 업데이트
- base.html의 메타 설명 수정으로 SEO 최적화

+++ vote/models.py
@@ -6,7 +6,7 @@ class School(models.Model):
     school_name = models.CharField(max_length=100)
     school_image = models.ImageField(upload_to='school_image', null=True)
     school_address = models.CharField(max_length=100, null=True)
-    school_score = models.BigIntegerField(default=0)
+    school_score = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     susi_school_pdf = models.FileField(upload_to="susi_school_pdf", null=True)
     jungsi_school_pdf = models.FileField(upload_to="jungsi_school_pdf", null=True)
